(#56) Refactor out the building of the using Stravaig... declaration
