+++ src/Stravaig.Extensions.Core.Analyzer.CodeFixes/SEC0001_UseStringHasContentAnalyzerCodeFixProvider.cs
@@ -3,15 +3,10 @@ using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.Rename;
-using Microsoft.CodeAnalysis.Text;
 using System;
-using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Composition;
-using System.Diagnostics;
 using System.Linq;
-using System.Reflection.Metadata;
 using System.Threading;
 using System.Threading.Tasks;
 using Document = Microsoft.CodeAnalysis.Document;
@@ -66,87 +61,11 @@ public class SEC0001_UseStringHasContentAnalyzerCodeFixProvider : CodeFixProvide
 
         rootNode = rootNode.ReplaceNode(stringIsNullOrWhiteSpaceExpression, hasContentExpression);
 
-        rootNode = UseStravaigExtensionsCore(rootNode);
+        rootNode = rootNode.UseStravaigExtensionsCore();
 
         return document.WithSyntaxRoot(rootNode);
     }
 
-    private CompilationUnitSyntax UseStravaigExtensionsCore(CompilationUnitSyntax oldRoot)
-    {
-        SyntaxList<UsingDirectiveSyntax> usingDeclarations;
-        if (oldRoot.Usings.Count != 0)
-        {
-            usingDeclarations = oldRoot.Usings;
-        }
-        else
-        {
-            var searchStart = GetNamespaceDeclaration(oldRoot);
-            usingDeclarations = searchStart?.Usings ?? oldRoot.Usings;
-        }
-
-        (bool usingExists, SyntaxNode insertBefore) = FindInsertionPoint(usingDeclarations);
-
-        if (usingExists)
-            return oldRoot;
-       
-        return insertBefore != null 
-            ? oldRoot.InsertNodesBefore(insertBefore, UsingStravaigExtensionsCore())
-            : oldRoot.AddUsings(UsingStravaigExtensionsCore());
-    }
-
-    private static BaseNamespaceDeclarationSyntax GetNamespaceDeclaration(CompilationUnitSyntax oldRoot)
-    {
-        
-        var childNodes = oldRoot.Members;
-        var fileScopesNamespaceNodes = childNodes.OfType<FileScopedNamespaceDeclarationSyntax>();
-        var namespaceDeclaration = fileScopesNamespaceNodes.FirstOrDefault();
-        if (namespaceDeclaration != null)
-            return namespaceDeclaration;
-
-        var traditionalNamespaceNodes = childNodes.OfType<NamespaceDeclarationSyntax>();
-        return traditionalNamespaceNodes.FirstOrDefault();
-    }
-
-    private static (bool, SyntaxNode) FindInsertionPoint(SyntaxList<UsingDirectiveSyntax> usings)
-    {
-        SyntaxNode insertBefore = null;
-        foreach (var usingDirective in usings)
-        {
-            string usingNamespace = "";
-            if (usingDirective.Name.Kind() == SyntaxKind.IdentifierName)
-            {
-                var identifierName = (IdentifierNameSyntax)usingDirective.Name;
-                usingNamespace = identifierName.Identifier.Text;
-            }
-            else if (usingDirective.Name.Kind() == SyntaxKind.QualifiedName)
-            {
-                var qualifiedIdentifierName = (QualifiedNameSyntax)usingDirective.Name;
-                usingNamespace = qualifiedIdentifierName.AsString();
-            }
-
-            if (usingNamespace.Equals("Stravaig.Extensions.Core"))
-                return (true, null);
-
-            if (usingNamespace.IsAfter("Stravaig.Extensions.Core", StringComparison.Ordinal))
-            {
-                insertBefore ??= usingDirective;
-            }
-        }
-
-        return (false, insertBefore);
-    }
-
-    private UsingDirectiveSyntax[] UsingStravaigExtensionsCore()
-    {
-        SimpleNameSyntax stravaig = SyntaxFactory.IdentifierName("Stravaig");
-        SimpleNameSyntax extensions = SyntaxFactory.IdentifierName("Extensions");
-        SimpleNameSyntax core = SyntaxFactory.IdentifierName("Core");
-        QualifiedNameSyntax stravaigExtensions = SyntaxFactory.QualifiedName(stravaig, extensions);
-        NameSyntax name = SyntaxFactory.QualifiedName(stravaigExtensions, core);
-        UsingDirectiveSyntax usingStatement = SyntaxFactory.UsingDirective(name);
-        return new[] { usingStatement };
-    }
-    
     private static InvocationExpressionSyntax BuildStringHasContentNodes(ExpressionSyntax stringExpression)
     {
         var simpleNameSyntax = SyntaxFactory.IdentifierName("HasContent");
